refactor:move function to stats controller
